import refacted : need test
